move ReplaceSelectedWith
